Use newline delimited list for image tags
An upstream change to the CSV parsing used by `docker/build-push-action`
caused the leading spaces in tag names to no longer be stripped.

This could be fixed by using the `>-` YAML multi-line string folder, but
`|` and a newline delimited list is clearer. This syntax matches what is
used by the `build-args` parameter to the `docker/build-push-action`
action.