Remove `.cargo/config.toml` snippets to use git CLI
Since https://github.com/artichoke/artichoke/issues/1826 has been fixed,
git clones happen much more quickly for Artichoke, so this workaround to
use the native git CLI is not necessary.

Additionally, the heredoc syntax feature of `Dockerfile`s that these
`COPY` instructions used requires syntax annotations [0] in the
`Dockerfile` sources that were not present.

[0] https://github.com/moby/buildkit/blob/86c33b66e176a6fc74b88d6f46798d3ec18e2e73/frontend/dockerfile/docs/syntax.md#here-documents

Per the docs:

> This feature is available since `docker/dockerfile:1.4.0` release.
>
>     # syntax=docker/dockerfile:1.4

These annotations are not present in the `Dockerfile`s in this
repository. I discovered this misconfiguration when attempting to build
with docker and buildx on Ubuntu in WSL. I'm not sure how this works in
CI.