Stop tagging Docker images with artichoke/artichoke git SHA
Stop tagging Docker images pushed to Docker Hub with the git SHA of the
source commit from the artichoke/artichoke repository.

This data is already tracked in OCI image metadata and is not useful for
pinning since the images in this repository are all nighlies.

This change removes the following image tag formats:

- `<git SHA>`
- `alpine-nightly-<git SHA>`
- `slim-nightly-<git SHA>`
- `ubuntu-nightly-<git SHA>`

This change will reduce the number of tags significantly and reduce the
number of active blobs stored in the repository, which will be useful as
the Artichoke organization shifts to a self-hosted registry.

See #141.