Ensure artichokeruby/artichoke:latest is always Ubuntu base image
On a recent pull to try out docker scout, I noticed that the :latest tag was a
Debian Bullseye base. It looks like all container builds get tagged as :latest
by the build matrix, leaving the :latest tag subject to a race condition.

Remove this erroneous tag to ensure :latest always points to an Ubuntu image.