Fix red squiggle in VSCode
`octokit/request-action` only defines a `route` input but it also will
slurp up all provided action inputs and interpolate them into the
provided route.

Because `owner` and `repo` aren't declared inputs to the action, VSCode
reports an error. Replace this with a hardcoded route to minimize the
chance of bugs and security vulnerabilities.